services/crud-django-service/pyproject.toml
[project]
name = "crud-django-service"
version = "0.1.0"
description = ""
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.11,<3.13"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.poetry]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
django = ">=5.2.5,<6.0.0"
djangorestframework = ">=3.16.1,<4.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
gunicorn = ">=23.0.0,<24.0.0"
celery = ">=5.5.3,<6.0.0"
django-celery-beat = ">=2.8.1,<3.0.0"
snowflake-id = "^1.0.2"
requests = "^2.32.4"
whitenoise = "^6.9.0"
djangorestframework-simplejwt = "^5.5.1"
pydantic = "^2.11.7"
pydantic-settings = "^2.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
ruff = "^0.5.3"
httpx = "^0.27.0"
ipython = "^9.4.0"
isort = "^6.0.1"
flake8 = "^7.3.0"
pytest-django = "^4.11.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
